requirements.txt
# Cross-platform dependencies
psutil>=5.9.0
tomli>=2.0.1; python_version < "3.11"

# macOS dependencies
pyobjc-core>=10.0,<12.0; sys_platform == "darwin"
pyobjc-framework-Cocoa>=10.0,<12.0; sys_platform == "darwin"

# Windows dependencies
pystray>=0.19.0; sys_platform == "win32"
Pillow>=10.0.0; sys_platform == "win32"
